requirements.txt
langchain>=0.2.10,<0.3.0
langchain-openai>=0.1.22
langchain-community>=0.2.10
langchain-text-splitters>=0.2.2
openai>=1.44.0
faiss-cpu>=1.8.0.post1
tiktoken>=0.7.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pypdf>=4.3.1
unstructured>=0.15.0
streamlit>=1.38.0
pytest>=8.3.2

